+++ packages/ice/src/createService.ts
@@ -23,13 +23,13 @@ import { generateRoutesInfo } from './routes.js';
 import * as config from './config.js';
 import { RUNTIME_TMP_DIR, WEB, RUNTIME_EXPORTS } from './constant.js';
 import createSpinner from './utils/createSpinner.js';
-import getRoutePaths from './utils/getRoutePaths.js';
 import ServerCompileTask from './utils/ServerCompileTask.js';
 import { getAppExportConfig, getRouteExportConfig } from './service/config.js';
 import renderExportsTemplate from './utils/renderExportsTemplate.js';
 import { getFileExports } from './service/analyze.js';
 import { getFileHash } from './utils/hash.js';
 import { logger } from './utils/logger.js';
+import RouteManifest from './utils/routeManifest.js';
 
 const require = createRequire(import.meta.url);
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -104,7 +104,7 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
   RUNTIME_EXPORTS.forEach(exports => {
     generatorAPI.addExport(exports);
   });
-
+  const routeManifest = new RouteManifest();
   const ctx = new Context<Config, ExtendsPluginAPI>({
     rootDir,
     command,
@@ -117,6 +117,8 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
         addEvent: addWatchEvent,
         removeEvent: removeWatchEvent,
       },
+      getRouteManifest: () => routeManifest.getNestedRoute(),
+      getFlattenRoutes: () => routeManifest.getFlattenRoute(),
       context: {
         webpack,
       },
@@ -161,6 +163,7 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
   const coreEnvKeys = getCoreEnvKeys();
 
   const routesInfo = await generateRoutesInfo(rootDir, routesConfig);
+  routeManifest.setRoutes(routesInfo.routes);
   const hasExportAppData = (await getFileExports({ rootDir, file: 'src/app' })).includes('dataLoader');
   const csr = !userConfig.ssr && !userConfig.ssg;
 
@@ -251,6 +254,7 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
       targetDir: RUNTIME_TMP_DIR,
       templateDir: coreTemplate,
       cache: dataCache,
+      routeManifest,
       ctx,
     }),
   );
@@ -263,11 +267,12 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
     run: async () => {
       try {
         if (command === 'start') {
-          const routePaths = getRoutePaths(routesInfo.routes)
+          const routePaths = routeManifest.getFlattenRoute()
             .sort((a, b) =>
               // Sort by length, shortest path first.
               a.split('/').filter(Boolean).length - b.split('/').filter(Boolean).length);
           return await start(ctx, {
+            routeManifest,
             taskConfigs,
             serverCompiler,
             getRoutesConfig,
@@ -280,6 +285,7 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
           });
         } else if (command === 'build') {
           return await build(ctx, {
+            routeManifest,
             getRoutesConfig,
             getDataloaderConfig,
             getAppConfig,